+++ App.js
@@ -1,7 +1,7 @@
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet, View, Text } from "react-native";
 import { useFonts } from "expo-font";
-import InitialPage from "./pages/InitialPage";
+import { InitialPage } from "./components/index";
 
 export default function App() {
   let [fontsLoaded] = useFonts({
