+++ components/index.js
@@ -1,0 +1,3 @@
+import InitialPage from "./initialPage/InitialPage";
+
+export { InitialPage };
