+++ components/initialPage/initialPage.styles.js
@@ -1,0 +1,25 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  background: {
+    flex: 1,
+  },
+  content: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  appTitle: {
+    fontSize: 40,
+    color: "white",
+    marginBottom: 8,
+    fontFamily: "lato",
+  },
+  appTagline: {
+    fontSize: 16,
+    color: "white",
+    fontFamily: "lato",
+  },
+});
+
+export default styles;
